Site Review and Tweaks

+++ app/views/users/partials/follow-form.blade.php
@@ -1,19 +1,21 @@
-@if ($user->isFollowedBy($me))
+@if ($signedIn)
+	@if ($user->isFollowedBy($me))
 
-	{{ Form::Open(['method' => 'DELETE', 'route' => ['follow_path', $user->id]]) }}
+		{{ Form::Open(['method' => 'DELETE', 'route' => ['follow_path', $user->id]]) }}
 
-		{{ Form::hidden('userIdToUnfollow', $user->id) }}
-		<button type="submit" class="btn btn-danger">Unfollow {{ $user->username }}</button>
+			{{ Form::hidden('userIdToUnfollow', $user->id) }}
+			<button type="submit" class="btn btn-danger">Unfollow {{ $user->username }}</button>
 
-	{{ Form::close() }}
+		{{ Form::close() }}
 
-@else
+	@else
 
-	{{ Form::Open(['route' => 'follows_path']) }}
+		{{ Form::Open(['route' => 'follows_path']) }}
 
-		{{ Form::hidden('userIdToFollow', $user->id) }}
-		<button type="submit" class="btn btn-primary">Follow {{ $user->username }}</button>
+			{{ Form::hidden('userIdToFollow', $user->id) }}
+			<button type="submit" class="btn btn-primary">Follow {{ $user->username }}</button>
 
-	{{ Form::close() }}
+		{{ Form::close() }}
 
+	@endif
 @endif
